move sight details geader delegate to widgets folder
