+++ lib/ui/screen/sight_details_screen/widgets/sight_details_header_delegate.dart
